drivers: Correct max_active parameter of some workqueues
This is probably the most uncertain commit I have authored.
Previously, by some famous commit which comes from myfluxi, some
workqueues are aded WQ_HIGHPRI flag. This is fine, but he changed
max_active argument as well without noticing it in commit message.
By default, the max_active parameter is 1 for create_workqueue(),
but he changed it to 0.

It is fine and good for performance in my opinion, as now a workqueue
are not bounded with a single worker. But as docs says,

"Some users depend on the strict execution ordering of ST wq. The
combination of@max_active of 1 and WQ_UNBOUND is used to achieve this
behavior.Work items on such wq are always queued to the unbound
worker-pools and only one work item can be active at any given time
thus achieving the same ordering property as ST wq." (https://
www.kernel.org/doc/html/v4.10/core-api/workqueue.html)

I am not quite sure about this, so recover them back. The difference
will not be noticeable, but better safe than sorry.

Signed-off-by: Wang Han <416810799@qq.com>