build: Mark-VI: Final build for N
I plan to start work on O soon, and do the bringup for Candy.

I'm really happy with this. Huge thanks to the developer of Kawori,
Sultanxda, Chad Cormier Russel, and Joe Maples for their guidance
and help. Stand alone, I'd just be a shitty developer so help is
always greatly appreciated.