arm64: configs: zippy_defconfig: Increase HZ to 1000
The way I see it, we keep pushing for tasks to become more realtime
and in doing so we now have a form of foreground scheduling.

100 hertz is not effective in this because it suffers from a delay
via RCU callbacks, and 300 hertz is lacking in the sense that using
a value not integrated by a power of 10 can lead to fragmentation
mostly seen in the scheduler.

By using a value of 1000 we do increase power consumption, and yes,
it is perceptible, however RT tasks are handled properly now, and
we have far greater interactivity only seen on newer devices such
as Marlin that make use of this value.

In the end, I feel it is worth it in the sense that tasks are issued
in the foreground properly now in next to an RT statuate.