mm: slub: Set default slub_max_order back to 3
That patch is rejected by upstream:

On Wed, May 11, 2011 at 01:38:47PM -0700, David Rientjes wrote:

Which is more reasonable? That an ordinary user gets a default that
is fairly safe even if benchmarks that demand the highest performance
from SLUB take a hit or that administrators running such workloads
set slub_max_order=3?